// apps/web/src/api/generated/model/userResource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DSH
 * OpenAPI spec version: 0.0.1
 */

export interface UserResource {
  id: string;
  name: string;
  email: string;
  avatarLink: string;
  /** @nullable */
  githubName: string | null;
  /** @nullable */
  githubLink: string | null;
  is_admin: boolean;
  /** @nullable */
  created_at: string | null;
}